Move MRV2 sparse offload metadata helper
